feat: add reviews API for product
