+++ package.json
@@ -27,6 +27,6 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "prisma": "^5.6.0"
+    "prisma": "latest"
   }
 }
