+++ server.js
@@ -10,6 +10,7 @@ import userRouter from './src/routes/user.js';
 import EventRouter from './src/routes/event.js';
 import RegisterEventRouter from './src/routes/register-event.js';
 import BannerRouter from './src/routes/banner.js';
+import reviewProductRouter from './src/routes/review-product.js';
 
 const app = express();
 
@@ -27,6 +28,7 @@ app.use('/api/auth/user', authRouter);
 app.use('/api/auth/customer', authCustomerRouter);
 app.use('/api/user', userRouter)
 app.use('/api/product', productRouter)
+app.use('/api/review', reviewProductRouter)
 app.use('/api/customer', customerRouter)
 app.use('/api/event', EventRouter)
 app.use('/api/event', RegisterEventRouter)
