+++ src/controllers/product/crud_product/create.js
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import multer from 'multer'
 import path from "path";
-import config from '../../config/app.js';
+import config from '../../../config/app.js';
 import sharp from 'sharp';
 import { PrismaClient } from '@prisma/client';
 import { v4 as uuidv4 } from 'uuid';
